+++ docs/CHANGELOG.md
@@ -115,6 +115,7 @@ Version 1.0.0
 - Implement new launcher GUI
 - Raise max FPS limit to 10000
 - Adjust default settings: linear pitch on, directinput on, vsync off, fps limit 240, resolution 1920x1080
+- Add functions to launch a test listen server on-demand: `levelm` console command, `-levelm` command line, and `Play (multi)` button in RED